Move Prefill and Decode Logic to utils
